+++ frontend/src/views/files/Editor.vue
@@ -50,7 +50,7 @@ export default {
   }),
   computed: {
     permissions() {
-      return getters.permissions();
+      return getters.sourcePermissions();
     },
     isDarkMode() {
       return getters.isDarkMode();